updated model options
