+++ src/components/picker/option/index.tsx
@@ -1,29 +1,5 @@
 import { Component, Element, Event, EventEmitter, h, Host, Method, Prop, State } from "@stencil/core"
-
-export interface Option {
-	element: HTMLSmoothlyPickerOptionElement
-	selected: boolean
-	readonly: boolean
-	visible: boolean
-	search: string[]
-	value: any
-	slotted: Node[]
-	set: {
-		selected: (selected: boolean) => void
-		visible: (visible: boolean) => void
-		readonly: (readonly: boolean) => void
-		search: (search: string[]) => void
-		value: (value: any) => void
-	}
-}
-export namespace Option {
-	export type Load = Omit<Option, "slotted">
-	export interface Created {
-		value: string
-		selected: boolean
-	}
-}
-
+import { Option } from "../../../model"
 @Component({
 	tag: "smoothly-picker-option",
 	styleUrl: "style.css",
